Restructure server directories
 - Separate routes
 - Separate middlewares
 - Rename library => libraries
