+++ packages/server/middlewares/cors.ts
@@ -1,0 +1,12 @@
+import type { CorsOptions } from 'cors';
+
+const whitelist = ['https://sunjae.kim'];
+export const corsOptions: CorsOptions = {
+  origin: (origin, callback) => {
+    if (!origin || whitelist.indexOf(origin) !== -1) {
+      callback(null, true);
+    } else {
+      callback(new Error('Not allowed by CORS'));
+    }
+  },
+};
